feat(wave): land NCCC-0731-ab7e — wave 2 of the pure-session near-miss crack (+216 net strict)
Wave: the remaining 310 functions of the [96,100) band (96.0-99.99), one Sonnet
agent per function, 80 Opus escalations, batched verify with a global
contested-file map. 157 fixed + 23 improved claims; 125 walls logged with
mechanisms to decomp.db ([NCCC-0731-ab7e/f<i>/<tier>] markers; DB pre-seeded
this time, so all 390 attempts logged as full prose - no journal backfill).

Landing: 174 patches (162 clean + 3-way merges + hand-unions). Notable
resolutions:
- Cache op-enum contradiction: f119s swapped kOpFileSize/kOpDirectory to crack
  ThreadDone but broke GetFileSizeAsync's retail li r10,1. The store pins
  kOpFileSize == 1; ThreadDone's case BODIES were the mislabeled half (retail's
  filesize op parks its out-pointer in mCacheDirList, consistent with
  GetFileSizeAsync's existing cast). Arm-body swap satisfies both at 100.
- f228s REVERTED: its RndPostProc mLuminanceMap layout change regressed the
  whole PostProc/PostProc_NG family (~33 fns); refuted by whole-TU measurement.
- Wind pair: f19o subsumes f107s (identical RndWind demotion, conservative
  variant kept). f80o-style duplicates and one corrupt inline patch
  regenerated from worktrees.
- MainHubPanel/f78s was byte-duplicate of an already-landed fix; SampleInst
  and Bitmap conflicts resolved to the measured first-lander.

A/B vs fresh pre-landing baseline: matched 41,415 -> 41,631 (+216),
masked_equal 1,487 -> 1,485 => honest floor +218, code% 35.70 -> 36.40.
Strict SET GAINED 259 / LOST 43, all 43 classified: 33 = target-symbol RENAME
churn (this build first materialized lane CC-8's 32 RTTI map corrections plus
this wave's repoints through the renamer), 7 = same symbol matched in a
different unit, 3 accepted & documented (2 map-label relocations at 0.00 whose
source donors are proven unchanged - map-correction follow-ups; 1 = Utl
aggregate static-block shift, fn_82438674 99.91, mechanism noted in-tree).

+++ src/system/bandobj/TrackPanelDir.cpp
@@ -620,7 +620,7 @@ void TrackPanelDir::Coda() {
             mEndingBonus->Start(!mTrackPanel->AutoVocals());
             for (int i = 0; i < mTracks.size(); i++) {
                 BandTrack *track = mTracks[i];
-                if (track)
+                if ((int)track)
                     track->EnterCoda();
             }
         }
